+++ NitroxClient/Resources/soundsWhitelist.csv
@@ -685,15 +685,15 @@ event:/sub/seamoth/impact_solid_soft;false;false;0
 event:/sub/seamoth/interior_loop;false;false;0
 event:/sub/seamoth/pulse;false;false;0
 event:/sub/seamoth/pulse_charge_loop;false;false;0
-event:/sub/seamoth/seaglide_equip;true;false;20
-event:/sub/seamoth/seaglide_light_off;true;false;10
-event:/sub/seamoth/seaglide_light_on;true;false;10
-event:/sub/seamoth/seaglide_loop_rpm;true;false;40
-event:/sub/seamoth/seaglide_rev_up;true;false;40
-event:/sub/seamoth/seamoth_light_off;false;false;0
-event:/sub/seamoth/seamoth_light_on;false;false;0
-event:/sub/seamoth/seamoth_loop_rpm;false;false;0
-event:/sub/seamoth/seamoth_rev_up;false;false;0
+event:/sub/seamoth/seaglide_equip;true;true;10
+event:/sub/seamoth/seaglide_light_off;true;true;7
+event:/sub/seamoth/seaglide_light_on;true;true;7
+event:/sub/seamoth/seaglide_loop_rpm;true;true;60
+event:/sub/seamoth/seaglide_rev_up;true;true;60
+event:/sub/seamoth/seamoth_light_off;true;true;10
+event:/sub/seamoth/seamoth_light_on;true;true;10
+event:/sub/seamoth/seamoth_loop_rpm;true;true;100
+event:/sub/seamoth/seamoth_rev_up;true;true;100
 event:/sub/seamoth/seamoth_stop;false;false;0
 event:/sub/seamoth/sonar_loop;false;false;0
 event:/sub/seamoth/storage_close;false;false;0
@@ -769,13 +769,13 @@ event:/tools/gravsphere/draw;false;false;0
 event:/tools/gravsphere/explode;false;false;0
 event:/tools/gravsphere/loop;false;false;0
 event:/tools/gravsphere/loop_actual;false;false;0
-event:/tools/knife/deploy;true;false;20
-event:/tools/knife/first_use;true;false;20
-event:/tools/knife/heat_hit;true;false;20
-event:/tools/knife/heat_idle;true;false;20
-event:/tools/knife/hit;true;false;20
-event:/tools/knife/swing;true;false;20
-event:/tools/knife/swing_surface;true;false;20
+event:/tools/knife/deploy;true;false;10
+event:/tools/knife/first_use;true;false;10
+event:/tools/knife/heat_hit;true;false;10
+event:/tools/knife/heat_idle;true;false;10
+event:/tools/knife/hit;true;false;10
+event:/tools/knife/swing;true;false;10
+event:/tools/knife/swing_surface;true;false;10
 event:/tools/lasercutter/cut_loop;true;false;20
 event:/tools/lasercutter/deploy;true;false;20
 event:/tools/lasercutter/door_cut;true;false;20
