+++ NitroxPatcher/Patches/Dynamic/FMODUWE_PlayOneShotImpl_Patch.cs
@@ -1,6 +1,7 @@
 ﻿using System.Reflection;
 using Harmony;
 using NitroxClient.GameLogic;
+using NitroxClient.GameLogic.FMOD;
 using NitroxModel.Core;
 using NitroxModel_Subnautica.DataStructures;
 using UnityEngine;
@@ -13,6 +14,11 @@ namespace NitroxPatcher.Patches.Dynamic
 
         private static readonly MethodInfo targetMethod = typeof(FMODUWE).GetMethod("PlayOneShotImpl", BindingFlags.NonPublic | BindingFlags.Static, null, new[] { typeof(string), typeof(Vector3), typeof(float) }, null);
 
+        public static bool Prefix()
+        {
+            return !NitroxModel.DataStructures.GameLogic.FMOD.FMODSuppressor.SuppressFMODEvents;
+        }
+
         public static void Postfix(string eventPath, Vector3 position, float volume)
         {
             if (fmodSystem.IsWhitelisted(eventPath, out bool isGlobal, out float radius))
@@ -24,7 +30,7 @@ namespace NitroxPatcher.Patches.Dynamic
         public override void Patch(HarmonyInstance harmony)
         {
             fmodSystem = NitroxServiceLocator.LocateService<FMODSystem>();
-            PatchPostfix(harmony, targetMethod);
+            PatchMultiple(harmony, targetMethod, true, true, false);
         }
     }
 }
